#526 #579: Revisar posibles problemas con tamaños de cadenas en el cliente; leer configuración para todos los discos.
git-svn-id: https://opengnsys.es/svn/branches/version1.0@3628 a21b9725-9963-47de-94b9-378ad31fedc9

+++ admin/Sources/Includes/ogAdmLib.h
@@ -129,7 +129,7 @@ typedef struct{		// Estructura de las tramas
 	char arroba;	// Caracter arroba siempre
 	char identificador[14];	// Identificador de la trama, siempre JMMLCAMDJ_MCDJ
 	char tipo;	// Tipo de mensaje
-	int lonprm; // Longitud en hexadecimal de los parámetros
+	long lonprm;	// Longitud en hexadecimal de los parámetros
 	char *parametros; // Parámetros de la trama
 }TRAMA;
 // ________________________________________________________________________________________________________
